Break the cycling schemas into their own module
